Improve rendering performance of query result panel
There was one big issue with the `useComunicaQuery` hook that provides
results to the `Results` panel: one of the dependencies of the hook was
the `results` array itself, which changed every single time the Comunica
stream updated.

What this meant in practice is that each time the Comunica stream had a
`data` event, all of the event handlers were detached, and then fresh
handlers were created and attached. When processing 10,000 records, that
would happen 10,000 times. This created a lot of overhead.

Removing the `results` dependency and changing how results were passed
to the provider (that is, removing the `immer` update and just using a
plain array), rendering 10,000 rows of a generic `?s ?p ?o` SPARQL query
went from taking ~60s to ~20s.

However, there's more performance to be had. For that same query,
rendering all results at once when the query is completed takes ~2.5s.
So, anything above that is overhead from rendering data immediately as
it's added.

What I did here is a pretty easy gain-- I throttled the function
updating the `result` state to 250ms. That means that `setResult` is
called, *at most*, once every 250ms.

(I made a special case for the first 100 results. There, `setResult`
isn't throttled, to allow initial data to show up immediately if it's
available).

Doing that throttling took rendering that same query down to ~5s. That's
still about a 2x overhead, but some re-rendering is necessary, and that
certainly beats a 25x overhead.