Remove some extraneous processing from RDFTable grid component
On every render of the RDFTable (e.g. when results are updated), each
binding was transformed from a Map object to a record of binding-term
pairs. Those objects-- created anonymously on the fly-- made it more
difficult for React Grid to check for equality. For a given row, the
same record in a result, even though structurally equal, but not
strictly equal. This made DOM reconciliation more intensive than it
needed to be.

This commit changes the RDFTable so that each row is simply represented
by the `Bindings` object returned by communica, which is identical
across renders. This speeds things up a bit.

Also fixes the types in RDFTable by taking advantage of React Grid's
generics.